Move "gui" package to "view"
