+++ src/main/java/de/htwsaar/prog3/carrental/controller/CustomerTableViewController.java
@@ -3,8 +3,8 @@ package de.htwsaar.prog3.carrental.controller;
 import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
-import de.htwsaar.prog3.carrental.gui.EditCustomerView;
-import de.htwsaar.prog3.carrental.gui.NewCustomerView;
+import de.htwsaar.prog3.carrental.view.EditCustomerView;
+import de.htwsaar.prog3.carrental.view.NewCustomerView;
 import de.htwsaar.prog3.carrental.model.Customer;
 import de.htwsaar.prog3.carrental.service.CustomerService;
 import de.htwsaar.prog3.carrental.util.GUIDialogUtil;
