+++ src/main/java/de/htwsaar/prog3/carrental/controller/EditCarViewController.java
@@ -1,6 +1,6 @@
 package de.htwsaar.prog3.carrental.controller;
 
-import de.htwsaar.prog3.carrental.gui.EditCarView;
+import de.htwsaar.prog3.carrental.view.EditCarView;
 import de.htwsaar.prog3.carrental.model.Car;
 import de.htwsaar.prog3.carrental.service.CarService;
 import de.htwsaar.prog3.carrental.util.i18n.I18nComponentsUtil;
