+++ src/main/java/de/htwsaar/prog3/carrental/controller/EmployeeTableViewController.java
@@ -3,8 +3,8 @@ package de.htwsaar.prog3.carrental.controller;
 import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
-import de.htwsaar.prog3.carrental.gui.EditEmployeeView;
-import de.htwsaar.prog3.carrental.gui.NewEmployeeView;
+import de.htwsaar.prog3.carrental.view.EditEmployeeView;
+import de.htwsaar.prog3.carrental.view.NewEmployeeView;
 import de.htwsaar.prog3.carrental.model.Employee;
 import de.htwsaar.prog3.carrental.service.EmployeeService;
 import de.htwsaar.prog3.carrental.util.GUIDialogUtil;
