+++ src/main/java/de/htwsaar/prog3/carrental/controller/NewCarViewController.java
@@ -1,6 +1,6 @@
 package de.htwsaar.prog3.carrental.controller;
 
-import de.htwsaar.prog3.carrental.gui.NewCarView;
+import de.htwsaar.prog3.carrental.view.NewCarView;
 import de.htwsaar.prog3.carrental.model.Car;
 import de.htwsaar.prog3.carrental.service.CarService;
 import de.htwsaar.prog3.carrental.util.i18n.I18nComponentsUtil;
