+++ src/main/java/de/htwsaar/prog3/carrental/controller/TableViewController.java
@@ -1,6 +1,6 @@
 package de.htwsaar.prog3.carrental.controller;
 
-import de.htwsaar.prog3.carrental.gui.CarTableView;
+import de.htwsaar.prog3.carrental.view.CarTableView;
 import de.htwsaar.prog3.carrental.util.EntityManagerUtil;
 import de.htwsaar.prog3.carrental.util.GUIDialogUtil;
 import de.htwsaar.prog3.carrental.util.i18n.I18nComponentsUtil;
