+++ src/main/java/de/htwsaar/prog3/carrental/view/EditEmployeeView.java
@@ -1,5 +1,6 @@
-package de.htwsaar.prog3.carrental.gui;
+package de.htwsaar.prog3.carrental.view;
 
+import de.htwsaar.prog3.carrental.model.Employee;
 import de.htwsaar.prog3.carrental.util.i18n.I18nComponentsUtil;
 import de.htwsaar.prog3.carrental.util.i18n.I18nStringsUtil;
 import de.htwsaar.prog3.carrental.util.i18n.I18nUtil;
@@ -10,42 +11,55 @@ import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 /**
- * Entry Point of the New Car Creation Dialog.
- *
+ * Dialog to edit an existing Employee
+ * 
  * @author Jens Thewes
+ *
  */
-public class NewCarView {
-    
+public class EditEmployeeView {
+
     private static Stage modalWindow;
+    private static Employee employee;
 
     /**
-     * Start the New Car Creation Dialog in a modal Window.
-     *
+     * Start the Edit Employee Dialog in a modal Window
+     * 
      * @param parentStage
      * @throws Exception
      */
-    public void start(Stage parentStage) throws Exception {
-        modalWindow  = new Stage();
-        // Load FXML document for the car configuration view wit the needed resource bundle
+    public void start(Stage parentStage, Employee employee) throws Exception {
+        modalWindow = new Stage();
+        EditEmployeeView.employee = employee;
+        // Load FXML document for the new employee view wit the needed resource bundle
         Parent scene =
-                FXMLLoader.load(getClass().getResource(I18nStringsUtil.getNewCarViewURL()),
+                FXMLLoader.load(getClass().getResource(I18nStringsUtil.getEditEmployeeViewURL()),
                         I18nUtil.getResourceBundleComponents());
         modalWindow.setTitle(I18nComponentsUtil.getStageTitleString());
         // Apply styling described in the FXML document
         modalWindow.setScene(new Scene(scene));
-        modalWindow.setMaxHeight(600);
-        modalWindow.setMaxWidth(900);
+        modalWindow.setMaxHeight(200);
+        modalWindow.setMaxWidth(600);
+        modalWindow.setHeight(200);
         // set the Owner of the modal window and the Modality of the new Stage (Modal Window)
         modalWindow.initOwner(parentStage);
         modalWindow.initModality(Modality.WINDOW_MODAL);
         modalWindow.show();
     }
-    
+
     /**
      * close the modal window
      * 
      */
     public static void closeModalWindow() {
         modalWindow.close();
+    }
+
+    /**
+     * get the car to edit
+     * 
+     * @return
+     */
+    public static Employee getEmployee() {
+        return employee;
     }
 }
